+++ integration_tests/shell.nix
@@ -4,8 +4,9 @@ pkgs.mkShell {
     (import ../. { inherit pkgs; })
     pkgs.go-ethereum
     pkgs.pystarport
+    pkgs.orchestrator
     pkgs.start-scripts
     pkgs.poetry
-    (pkgs.poetry2nix.mkPoetryEnv { projectDir = ./.; })
+    (import ../nix/testenv.nix { inherit pkgs; })
   ];
 }